// src/app/models/Reporte.model.ts
export class ReporteExcel {
    NombreBuscado: string;
    NombreCoincidencia: string;
    CodigoRegistro: string;
    SenialLAFT: string;
    CodPorducto: string;
    producto: string;
    NroPoliza: number;
    Certificado: number;
    TipoCliente: string;
    CodCliente: string;
    TipoDocumento: string;
    NroDocumento: string;
    NombresCompletos: string;
    DepartamentoDireccion: string;
    FechaInicioVigenciaPoliza: string;
    FechaFinVigenciaPoliza: string;
    FechaInicioVigenciaCertificado: string;
    FechaFinVigenciaCertificado: string;
    FechaAnulacion: string;
}