// src/app/models/PolicyRegister.ts
import { DecimalPipe } from '@angular/common';

export class PolicyRegister {

    codProduct: string;
    desProduct: string;
    nroPolicy: number;
    nroCertificate: number;
    rol: string;
    codClient: string;
    documento: string;
    nroDocumento: string;
    nombre: string;
    direccion: string;
    fechaInicioVigenciaPol: string;
    fechaFinVigenciaPol: string;
    fechaInicioVigenciaCert: string;
    fechaFinVigenciaCert: string;
    fechaAnulacion: string;
    ultimaPrima: number;
    fechaUltimaPagada: string;
    placa: string;
    tipo: string;


    nameSearch: string;
};