// src/app/models/sbsReport.model.ts
export class SbsReport {
    id: string;
    fechaInicio: string;
    fechaFin: string;
    tipoCambio: number;
    tipoOper: string;
    monto: string;
    usuario: string;
    estado: string;
    fecha: string;
    fechaIniProc: string;
    fechaFinProc: string;
    nombreReporte: string;
    desReporte: string;
}